move the completion command creation into its own package
